[refactor] 고객 파트 모듈 분리를 위한 패키지 구조 변경
